Update guest guard component
