+++ login-workflow/example/src/screens/Login.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { LoginScreen } from '@brightlayer-ui/react-auth-workflow';
 import EatonLogo from '../assets/images/eaton_stacked_logo.png';
 import { DebugComponent } from '../components/DebugComponent';
-import { useLocation } from 'react-router-dom';
 
 export const Login = (): JSX.Element => (
     // Everything on the login screen should work out of the box without having to specify ANY overrides for props here
