+++ login-workflow/example/src/screens/index.ts
@@ -1,4 +1,3 @@
-export * from './AppRouter';
 export * from './ExampleHome';
 export * from './Login';
-export * from './PageNotFound';
+export * from './CustomPage';
