+++ login-workflow/src/new-architecture/components/index.ts
@@ -5,3 +5,4 @@ export * from './WorkflowFinishState';
 export * from './RegistrationWorkflow';
 export * from './SetPassword';
 export * from './ChangePasswordDialog';
+export * from './Guards';
