--- login-workflow/src/new-architecture/experimental/index.ts
@@ -1,1 +1,0 @@
-export * from './ReactAuthGuard';
