+++ login-workflow/src/new-architecture/index.ts
@@ -1,6 +1,5 @@
 export * from './components';
 export * from './contexts';
-export * from './experimental';
 export * from './screens';
 export * from './utils';
 export * from './constants';
